refactor(tests): split the CLAUDE.md guards by the question each one answers
The Stop-gate caught a crossing this run caused: test_claude_md_template.py had
gone 160 -> 446 against a 300 limit, with no baseline entry. That file owns one
question -- do the greenfield template and the brownfield render stay mirrored
on the onboarding bullets -- and this iterate bolted a second, larger question
onto it. Split along the seam that was already there:

  test_claude_md_template.py                     160  the original mirror
  test_claude_md_standing_request.py             207  the grant's CONTENT
  ..adopt/tests/..standing_request_append.py     176  DELIVERY via the writer

The delivery half moves to adopt's own test root. That is its proper home by
the by-artifact convention, and it drops the subprocess trampoline: inside that
root `from lib.artifact_writer import write_claude_md` resolves directly, so
those tests drive the writer in-process instead of shelling out.

Coverage went up, not sideways -- 15 tests in one file became 18 across three.
The grant assertion that had been a loop inside the mirror test is now a test
of its own, and a new guard asserts the adopt-side markers still exist in the
shipped constant, because a marker tuple that drifts from the constant asserts
nothing at all.

shared/tests 6294 -3 (the moved tests), adopt 597 +6, iterate 510, integration
429, ruff clean, F0.5 re-run at 31 tests exit 0. Spec 3e and the ledger record
the split; rows repointed at the files that now hold each behavior.

Run-ID: iterate-2026-07-28-review-subagents-standing-request
